# api/openapi.yaml
openapi: 3.0.3
info:
  title: iRODS REST API
  version: 0.1.0
  description: |
    REST-ful iRODS running on GoLang.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/principal:
    get:
      operationId: getPrincipals
      tags: [user]
      summary: Search iRODS users and groups
      description: |
        Searches user and group principals by name prefix for combined selector
        UIs. Use the separate `/api/v1/user` and `/api/v1/usergroup` prefix
        searches when a combined result set is not needed.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: query
          required: true
          schema:
            type: string
            minLength: 3
          description: Principal name prefix to search.
        - in: query
          name: kind
          required: false
          schema:
            type: array
            items:
              type: string
              enum: [user, group]
          style: form
          explode: true
          description: Optional principal kind filter. May be repeated.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum result count.
      responses:
        "200":
          description: Matching principals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrincipalSearchResponse"
        "400":
          description: Invalid search option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while searching principals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user:
    get:
      operationId: getUsers
      tags: [user]
      summary: List iRODS users
      description: |
        Lists visible iRODS user principals in the configured zone. When `prefix`
        is supplied for autocomplete-style search, it must be at least 3 characters.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [rodsuser, rodsadmin]
          description: Optional user type filter.
        - in: query
          name: prefix
          required: false
          schema:
            type: string
            minLength: 3
          description: Optional username prefix filter for autocomplete.
      responses:
        "200":
          description: Users visible in the zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCollectionResponse"
        "400":
          description: Invalid user list option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postUser
      tags: [user]
      summary: Create one iRODS user
      description: |
        Requires the authenticated iRODS user to have type `rodsadmin` or `groupadmin`.
        Synchronization clients may use `reconcile=true` to make creation repeatable
        when the existing principal is already the requested user type. The operation
        remains a conflict when the existing principal is a group or when the existing
        user type differs from the requested type.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: reconcile
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Desired-state sync mode. Treat an existing user with the requested type
            as success instead of returning 409. A type mismatch or non-user
            principal with the same name remains a conflict. If a password is supplied
            for an existing matching user, the password is applied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "200":
          description: Desired user state satisfied when reconciliation mode is used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid create request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Authenticated iRODS user is not rodsadmin or groupadmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/user/{user_name}/avu:
    get:
      operationId: getUserAVUs
      tags: [user]
      summary: List AVUs for an iRODS user
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: attrib
          required: false
          schema:
            type: string
          description: Exact AVU attribute name to include in the response.
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [id, attrib, value, unit, created_at, updated_at]
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: User AVUs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAVUResponse"
        "400":
          description: Invalid AVU list option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while reading user AVUs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postUserAVU
      tags: [user]
      summary: Add a single AVU to an iRODS user
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathAVUCreateRequest"
      responses:
        "201":
          description: User AVU created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAVUSingleResponse"
        "400":
          description: Invalid AVU request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while creating user AVU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user/{user_name}/avu/{avu_id}:
    put:
      operationId: putUserAVU
      tags: [user]
      summary: Update a single AVU on an iRODS user
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: path
          name: avu_id
          required: true
          schema:
            type: string
          description: AVU identifier returned by the user AVU listing endpoint.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathAVUUpdateRequest"
      responses:
        "200":
          description: User AVU updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAVUSingleResponse"
        "400":
          description: Invalid AVU request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User or AVU not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while updating user AVU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteUserAVU
      tags: [user]
      summary: Delete a single AVU from an iRODS user
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: path
          name: avu_id
          required: true
          schema:
            type: string
          description: AVU identifier returned by the user AVU listing endpoint.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
      responses:
        "204":
          description: User AVU deleted
        "400":
          description: Invalid AVU identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User or AVU not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while deleting user AVU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/usergroup:
    get:
      operationId: getUserGroups
      tags: [user]
      summary: List iRODS groups
      description: |
        Lists visible iRODS groups in the configured zone. When `prefix`
        is supplied for autocomplete-style search, it must be at least 3 characters.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: prefix
          required: false
          schema:
            type: string
            minLength: 3
          description: Optional group-name prefix filter for autocomplete.
      responses:
        "200":
          description: Groups visible in the zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupCollectionResponse"
        "400":
          description: Invalid group list option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postUserGroup
      tags: [user]
      summary: Create one iRODS group
      description: |
        Generic REST surface for creating iRODS groups. Workflows that
        need to provision iRODS groups should call this route rather than an
        `/api/v1/ext` route.

        Requires the authenticated iRODS user to have type `rodsadmin` or `groupadmin`.
        Synchronization clients may use `reconcile=true` to make group creation
        repeatable when the existing principal is already an iRODS group. Created
        groups are marked with `iRODS:USER_SYNCH:MANAGED=true`.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: reconcile
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Desired-state sync mode. Treat an existing group as success instead of
            returning 409. A same-name non-group principal remains a conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserGroupCreateRequest"
      responses:
        "200":
          description: Desired group state satisfied when reconciliation mode is used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupResponse"
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupResponse"
        "400":
          description: Invalid create request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Group already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Authenticated iRODS user is not rodsadmin or groupadmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/usergroup/{group_name}/avu/{avu_id}:
    put:
      operationId: putUserGroupAVU
      tags: [user]
      summary: Update a single AVU on an iRODS group
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: group_name
          required: true
          schema:
            type: string
          description: iRODS group name.
        - in: path
          name: avu_id
          required: true
          schema:
            type: string
          description: AVU identifier returned by the group AVU listing endpoint.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathAVUUpdateRequest"
      responses:
        "200":
          description: Group AVU updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupAVUSingleResponse"
        "400":
          description: Invalid AVU request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Group or AVU not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while updating group AVU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteUserGroupAVU
      tags: [user]
      summary: Delete a single AVU from an iRODS group
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: group_name
          required: true
          schema:
            type: string
          description: iRODS group name.
        - in: path
          name: avu_id
          required: true
          schema:
            type: string
          description: AVU identifier returned by the group AVU listing endpoint.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
      responses:
        "204":
          description: Group AVU deleted
        "400":
          description: Invalid AVU identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Group or AVU not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while deleting group AVU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      scheme: basic
  schemas:
...
    UserAVUResponse:
      type: object
      required: [user_name, zone, avus]
      properties:
        user_name:
          type: string
        zone:
          type: string
        links:
          type: object
          properties:
            self:
              $ref: "#/components/schemas/ActionLink"
            create:
              $ref: "#/components/schemas/ActionLink"
        avus:
          type: array
          items:
            $ref: "#/components/schemas/AVUMetadata"
        count:
          type: integer
          description: Number of AVUs returned after filtering and pagination.
        total:
          type: integer
          description: Number of AVUs matching the filter before pagination.
        offset:
          type: integer
          description: Number of matching AVUs skipped before returning results.
        limit:
          type: integer
          description: Requested page size, or 0 when no limit was requested.
...
    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    UserCollectionResponse:
      type: object
      required: [users, count, zone, links]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        count:
          type: integer
        zone:
          type: string
        type:
          type: string
          enum: [rodsuser, rodsadmin, groupadmin]
        prefix:
          type: string
        links:
          type: object
          properties:
            self:
              $ref: "#/components/schemas/ActionLink"
            create:
              $ref: "#/components/schemas/ActionLink"
...
    UserCreateRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          description: iRODS username to create.
        type:
          type: string
          enum: [rodsuser, rodsadmin]
          description: iRODS user type for the new user.
        password:
          type: string
          format: password
          description: Optional initial iRODS user password.
...
    PrincipalSearchResponse:
      type: object
      required: [principals, count, zone, query]
      properties:
        principals:
          type: array
          items:
            $ref: "#/components/schemas/PrincipalSearchResult"
        count:
          type: integer
        zone:
          type: string
        query:
          type: string
        kinds:
          type: array
          items:
            type: string
            enum: [user, group]
        limit:
          type: integer
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: object
          description: Optional field-level validation messages keyed by request field name.
          additionalProperties:
            type: string